Refactor task handling, screenshot capture, and improve OCR exception handling
- Change `idTpDailyTask` type from `Long` to `Integer`
- Use `Platform.runLater` for clearing logs in
`ConsoleLogLayoutController`
- Comment out `captureScrenshotViaADB` in `LauncherLayoutController`
- Add `restartAdbServer` method in `EmulatorManager`
- Refactor `captureScreenshotViaADB` to return `ByteArrayInputStream`
- Update `ServScheduler` to use `DelayedTask` directly in task
scheduling
- Add `IOException` and `TesseractException` handling in OCR methods
- Replace `TASK_NAME` with `taskName` in `InitializeTask`
- Add `restartAdbServer` call in `InitializeTask` after 5 attempts
- Add `buscarTemplate` method to handle `InputStream` in
`ImageSearchUtil`

+++ wos-ot/src/main/java/cl/camodev/wosbot/ot/DTODailyTaskStatus.java
@@ -5,14 +5,14 @@ import java.time.LocalDateTime;
 public class DTODailyTaskStatus {
 
 	private Long idProfile;
-	private Long idTpDailyTask;
+	private Integer idTpDailyTask;
 	private LocalDateTime lastExecution;
 	private LocalDateTime nextSchedule;
 
 	public DTODailyTaskStatus() {
 	}
 
-	public DTODailyTaskStatus(Long idProfile, Long idTpDailyTask, LocalDateTime lastExecution, LocalDateTime nextSchedule) {
+	public DTODailyTaskStatus(Long idProfile, Integer idTpDailyTask, LocalDateTime lastExecution, LocalDateTime nextSchedule) {
 		this.lastExecution = lastExecution;
 		this.nextSchedule = nextSchedule;
 		this.idProfile = idProfile;
@@ -27,11 +27,11 @@ public class DTODailyTaskStatus {
 		this.idProfile = idProfile;
 	}
 
-	public Long getIdTpDailyTask() {
+	public Integer getIdTpDailyTask() {
 		return idTpDailyTask;
 	}
 
-	public void setIdTpDailyTask(Long idTpDailyTask) {
+	public void setIdTpDailyTask(Integer idTpDailyTask) {
 		this.idTpDailyTask = idTpDailyTask;
 	}
